Add files via upload
If you want to use this file, you gotta create an .inp file and an .out file. The input file should have the number of rows, the number of columns. In the maze, A is the beginning, B is the ending, * is the valid path, # is the wall. The .out file should have the + path as a found path.